StoreBrand listing   -feature and priority wise   -non feature and priority wise  Updating feature and priority of store brands

+++ src/main/java/com/yetistep/delivr/dao/inf/ActionLogDaoService.java
@@ -16,5 +16,5 @@ import java.util.List;
 public interface ActionLogDaoService extends GenericDaoService<Long, ActionLogEntity> {
      public List<ActionLogEntity> findActionLogPaginated(Page page) throws Exception;
 
-     public Long getTotalNumberOfActionLogs() throws Exception;
+     public Integer getTotalNumberOfActionLogs() throws Exception;
 }
